refactor: 불필요한 클래스 제거 및 RecentlyViewedProductRemoteRepository 리팩터링
