--- app/src/main/java/woowacourse/shopping/data/entity/RecentlyViewedProductEntity.kt
@@ -1,7 +1,0 @@
-package woowacourse.shopping.data.entity
-
-data class RecentlyViewedProductEntity(
-    val id: Long,
-    val productId: Long,
-    val viewedDateTime: String
-)
