+++ app/src/main/java/woowacourse/shopping/ui/productdetail/ProductDetailActivity.kt
@@ -9,18 +9,14 @@ import androidx.appcompat.app.AppCompatActivity
 import com.bumptech.glide.Glide
 import woowacourse.shopping.R
 import woowacourse.shopping.data.cart.CartItemRemoteRepository
-import woowacourse.shopping.data.cart.CartItemRemoteService
 import woowacourse.shopping.data.database.DbHelper
 import woowacourse.shopping.data.product.ProductRemoteRepository
-import woowacourse.shopping.data.product.ProductRemoteService
-import woowacourse.shopping.data.recentlyviewedproduct.RecentlyViewedProductMemoryDao
-import woowacourse.shopping.data.recentlyviewedproduct.RecentlyViewedProductRepositoryImpl
+import woowacourse.shopping.data.recentlyviewedproduct.RecentlyViewedProductRemoteRepository
 import woowacourse.shopping.databinding.ActivityProductDetailBinding
 import woowacourse.shopping.ui.cart.CartActivity
 import woowacourse.shopping.ui.productdetail.uistate.LastViewedProductUIState
 import woowacourse.shopping.ui.productdetail.uistate.ProductDetailUIState
 import woowacourse.shopping.utils.PRICE_FORMAT
-import woowacourse.shopping.utils.ServerConfiguration
 import woowacourse.shopping.utils.customview.AddToCartDialog
 
 class ProductDetailActivity : AppCompatActivity(), ProductDetailContract.View {
@@ -29,16 +25,10 @@ class ProductDetailActivity : AppCompatActivity(), ProductDetailContract.View {
     }
     private val presenter: ProductDetailContract.Presenter by lazy {
         ProductDetailPresenter(
-            this, ProductRemoteRepository(),
-            CartItemRemoteRepository(
-                CartItemRemoteService(ServerConfiguration.host)
-            ),
-            RecentlyViewedProductRepositoryImpl(
-                RecentlyViewedProductMemoryDao(
-                    DbHelper.getDbInstance(this)
-                ),
-                ProductRemoteService(ServerConfiguration.host)
-            )
+            this,
+            ProductRemoteRepository(),
+            CartItemRemoteRepository(),
+            RecentlyViewedProductRemoteRepository(DbHelper.getDbInstance(this))
         )
     }
     private val lastViewedProductViewHolder: LastViewedProductViewHolder by lazy {
